推荐系统基本完成，异构数据融合上未完成

+++ recommender/ItemCFRecommender/src/main/scala/pers/xxl/itemcf/ItemCFRecommender.scala
@@ -3,7 +3,7 @@ package pers.xxl.itemcf
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.SparkSession
 
-case class ProductRating( userId: Int, productId: Int, score: Double, timestamp: Int )
+case class ProductRating( userId: Int, productId: Int, score: Double, timestamp: BigInt )
 case class MongoConfig( uri: String, db: String )
 
 // 定义标准推荐对象
